+++ extras/integration_tests/web_wasm/test/drift_wasm_test.dart
@@ -142,7 +142,7 @@ void main() {
           }
 
           group(
-            'initialization from ',
+            'initialization from',
             () {
               test('static blob', () async {
                 await driver.enableInitialization(InitializationMode.loadAsset);