+++ extras/integration_tests/web_wasm/web/worker.dart
@@ -1,5 +1,12 @@
 import 'package:drift/wasm.dart';
+import 'package:sqlite3/wasm.dart';
 
 void main() {
-  WasmDatabase.workerMainForOpen();
+  WasmDatabase.workerMainForOpen(setupAllDatabases: (db) {
+    db.createFunction(
+      functionName: 'database_host',
+      function: (args) => 'worker',
+      argumentCount: const AllowedArgumentCount(1),
+    );
+  });
 }